WIP: corrected float32 label enc cast
Former-commit-id: 8f322ff10ff23436bfc5ea0dc9001ff05d061c6b [formerly 8f322ff10ff23436bfc5ea0dc9001ff05d061c6b [formerly ac050ef097e84a12108a32a279231fa8487b1ff1]]
Former-commit-id: b289a31c691e99cdf6b1f34ebac3dada25b9ab13
Former-commit-id: 39f8d1fffd245cc614feaa560bd8d14b623a832f

+++ arn/scripts/research/clip/kinetics_clip_img_encode.py
@@ -193,12 +193,12 @@ def main(
     # baseline. Save or reuse whatever boilerplate is around CLIP encoding here
 
     if load_encoded_labels and label_path:
-        encoded_labels = torch.load(label_path)
+        encoded_labels = torch.load(label_path).type(torch.float32)
     else:
         encoded_labels = None
 
     if load_encoded_images and image_path:
-        encoded_images = torch.load(image_path).type(torch.float32)
+        encoded_images = torch.load(image_path)
     else:
         encoded_images = None
         video_paths = []
